// apps/web/src/lib/api/generated/main/model/create-watched-item-quick-add-dto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export interface CreateWatchedItemQuickAddDto {
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  itemId: number;
  /**
     * @minLength 1
     * @maxLength 255
     */
  itemName: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  world: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  guildId: string;
}