// apps/web/src/lib/api/generated/main/model/watched-item-response-dto-output.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { NotificationRuleResponseDtoOutput } from './notification-rule-response-dto-output';
import type { WatchedItemSnapshotResponseDto } from './watched-item-snapshot-response-dto';

export interface WatchedItemResponseDtoOutput {
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  id: number;
  userId: string;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  itemId: number;
  itemName: string;
  world: string;
  enabled: boolean;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     * @nullable
     */
  notificationRuleId: number | null;
  createdAt: string;
  updatedAt: string;
  itemSnapshot: WatchedItemSnapshotResponseDto | null;
  notificationRule: NotificationRuleResponseDtoOutput | null;
}